requirements.txt
websockets==12.0
aiohttp==3.9.3
numpy==1.26.4
python-dotenv==1.0.1
